backend_requirement.txt
flask
requests
firebase_admin
python-dotenv